Added folders in validations to allow easier navigation and moved the validation files to correct folder
