+++ Backend/Validation/FollowedValidations/FollowedValidation.cs
@@ -1,7 +1,7 @@
 using Backend.Models;
 using FluentValidation;
 
-namespace Backend.Validation;
+namespace Backend.Validation.FollowedValidations;
 
 public class FollowedValidation : AbstractValidator<Followed>
 {
