+++ Backend/Validation/PostItemValidations/PostItemValidation.cs
@@ -1,7 +1,7 @@
 using Backend.Models;
 using FluentValidation;
 
-namespace Backend.Validation;
+namespace Backend.Validation.PostItemValidations;
 
 public class PostItemValidation : AbstractValidator<PostItem>
 {
